cli: deprecate `jj checkout`
Summary: As discussed in Discord, on GitHub, and elsewhere, this change
deprecates the use of `jj checkout` and suggests users use `jj new` exclusively
instead.

To further drive the bit home, by default, `jj checkout` (and `jj co`) is now
hidden. This will hopefully stop new users from running into it.

It is planned to remove `jj checkout` in version 0.18, which would give a 4
month grace period for users to adapt. After that point, the warnings will
likely remain for a long time, as there's little cost to removing it.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: I7a1adfc9168fce1f25cf5d4c4c313304769e41a1