cli: deprecate `jj merge`
Summary: As discussed in Discord, on GitHub, and elsewhere, this change
deprecates the use of `jj merge` and suggests users use `jj new` exclusively
instead.

To further drive the bit home, by default, `jj merge` is now hidden. This will
hopefully stop new users from running into it.

It is planned to remove `jj merge` in version 0.18, which would give a 4 month
grace period for users to adapt. After that point, the warnings will likely
remain for a long time, as there's little cost to removing it.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: I94938aca9d3e2aa12d1394a5fbc58acce3185b56